Catch the TypeErrors encountered by the pdf_parser module
The pdf parser module sometimes struggles to parse a PDF, either because
it is password locked, encrypted, too big, malformed, etc.

If an error is encountered, Beautiful soup tries to load a
malformed/empty file, which leads to a NoneType has no iterable or TypeError.

To fix this issue, we're catching the exception and log it as a warning.

This should fix issue #125 and #126