Refactor split and parse to work one reference at a time
Since now we had to pass along information about the reference id and
document id more than it was neccesary. This final commit allow us to
run the split and parse without information about the document or
reference while being able to reconstruct the information outside in
refparse in order to store in the document. This means that no reference
id is needed inside parse and it also means parse works one reference at
a time which makes it easier to expose to the user. As a happy side
effect performance seems to have improved 20x.