Move all code under policytool/
Since joining the refparse and wsf-web-scraper repos, we've had two
Python installs and docker images running around: one for refparse and
one for everythin gelse. This finaly consolidates us down to a single
directory, along the lines of a typical Python project (where you have a
repo of a given name, foo, and a directory within it, foo/, for all your
module code).

Highlights include:

- moving all refparse code into policytool/refparse
- moving the new algo_evaluation code into policytool/algo_evaluation
- consolidating all of policytool/Makefile into the top-level Makefile
- consolidating the top-level and policytool/requirements.txt into one
  (some stuff we don't use was dropped from the top-level one)
- deleting the refparse model_tests (because Liz says we can!)
- getting all tests to pass
- updating the docker builds for both the base and main image
  (now named policytool, for a short time only!)
- updating the web static build so it builds in the toplevel build
  dir
- moving docker-compose.yml to the top level and updating it accordingly

Remainders:

- consolidate policytool/README.md -> README.md