policytool: make airflow scrapy tasks fail
Issue: https://github.com/wellcometrust/policytool/issues/192

While testing SSO credentials it became clear that scrapes are
running "to completion" even when the pipeline fails to complete,
or when items failed to be saved. Fix this by adding Scrapy signal
handlers to our airflow task, and by firing an additional signal
if we failed to write the final manifest file at the end of the
pipeline.

Other changes:

* Fix duplicate (and verbose -- boto debug logs were getting thrown in)
  logging by making our Crawler not add a root handler.
* Log scrapy settings on a single line; it was super long otherwise.
* Report exceptions in our blocking thread to Sentry, since
  they won't show up in the Twisted error log.
* Note that our item pipeline's uploader is blocking (!).